A port of ghodss/yaml#39