@dims 2021-08-14
Hey @kevin-hanselman thanks for doing this. Could i persuade you to consider adding some benchmarks? (the linked PR hints at "memory footprint" gains ... so `-benchmem`?)