@kevin-hanselman 2021-08-15
Added a (very) basic benchmark test. Here's the results on my machine:

This branch:
```
goos: linux
goarch: amd64
pkg: sigs.k8s.io/yaml
cpu: Intel(R) Core(TM) i3-7100 CPU @ 3.90GHz
BenchmarkMarshal-4        153171              7745 ns/op            9530 B/op         50 allocs/op
BenchmarkUnmarshal-4       36927             32184 ns/op           11413 B/op        151 allocs/op
PASS
ok      sigs.k8s.io/yaml        2.788s
```

master:
```
goos: linux
goarch: amd64
pkg: sigs.k8s.io/yaml
cpu: Intel(R) Core(TM) i3-7100 CPU @ 3.90GHz
BenchmarkMarshal-4        249537              4672 ns/op            8937 B/op         42 allocs/op
BenchmarkUnmarshal-4       65419             18461 ns/op           10891 B/op        141 allocs/op
PASS
ok      sigs.k8s.io/yaml        2.615s
```

I'll need to think more about these results and code before drawing any conclusions.