@dims 2021-09-03
@deads2k @sttts @deads2k thoughts please!

cc @kubernetes-sigs/kubernetes-sig-api-machinery-reviewers