@VictorLowther 2021-09-07
On Tue, Sep 7, 2021, 16:11 Daniel Smith ***@***.***> wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In yaml.go
> <https://github.com/kubernetes-sigs/yaml/pull/59#discussion_r703833418>:
>
> > +}
> +
> +// Encode serializes a value into YAML and writes it using the Encoder's
> +// io.Writer.
> +func (e *Encoder) Encode(o interface{}) error {
> +	pipeReader, pipeWriter := io.Pipe()
> +	defer pipeReader.Close()
> +
> +	var intermediate interface{}
> +	jsonEnc := json.NewEncoder(pipeWriter)
> +	// We are using yaml.Decoder here (instead of json.Decoder) because the Go
> +	// JSON library doesn't try to pick the right number type (int, float,
> +	// etc.) when unmarshalling to interface{}; it always picks float64.
> +	// go-yaml preserves the number type.
> +	yamlDec := yaml.NewDecoder(pipeReader)
> +	yamlEnc := yaml.NewEncoder(e.w)
>
> This is doing an encode/decode/encode? Why? There has to be a better way
> than this? There should be some comment explaining the purpose?
>

Better is a fuzzy metric. I do not use this package in any performance
critical paths, it got the job done, and I have never had to touch it again.

—
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes-sigs/yaml/pull/59#pullrequestreview-748430367>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAF4EEXZBEZTWVREDX6D3NDUAZ5W3ANCNFSM5CFGWM3Q>
> .
> Triage notifications on the go with GitHub Mobile for iOS
> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
> or Android
> <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>
>