@luxas 2021-09-23
Cross-referencing https://github.com/kubernetes-sigs/yaml/pull/61#issuecomment-920711786 with information about potential blockers / info to think about with regards to this PR.